Refactor the integrations module organization #350
