+++ workflow/integrations/__init__.py
@@ -6,69 +6,31 @@
 # See https://aboutcode.org for more information about AboutCode FOSS projects.
 #
 
-from django.conf import settings
+from workflow.integrations.base import BaseIntegration
+from workflow.integrations.github import GitHubIntegration
+from workflow.integrations.gitlab import GitLabIntegration
+from workflow.integrations.jira import JiraIntegration
 
-import requests
+__all__ = [
+    "BaseIntegration",
+    "GitHubIntegration",
+    "GitLabIntegration",
+    "JiraIntegration",
+]
 
-DEJACODE_SITE_URL = settings.SITE_URL.rstrip("/")
+
+def get_class_for_tracker(issue_tracker_id):
+    if "github.com" in issue_tracker_id:
+        return GitHubIntegration
+    elif "gitlab.com" in issue_tracker_id:
+        return GitLabIntegration
+    elif "atlassian.net" in issue_tracker_id:
+        return JiraIntegration
 
 
-class BaseIntegration:
-    """Base class for managing issue tracker integrations from DejaCode requests."""
-
-    default_timeout = 10
-
-    def __init__(self, dataspace):
-        if not dataspace:
-            raise ValueError("Dataspace must be provided.")
-        self.dataspace = dataspace
-        self.session = self.get_session()
-
-    def get_session(self):
-        session = requests.Session()
-        session.headers.update(self.get_headers())
-        return session
-
-    def get_headers(self):
-        """
-        Return authentication headers specific to the integration.
-        Must be implemented in subclasses.
-        """
-        raise NotImplementedError
-
-    @staticmethod
-    def make_issue_title(request):
-        return f"[DEJACODE] {request.title}"
-
-    @staticmethod
-    def make_issue_body(request):
-        request_url = f"{DEJACODE_SITE_URL}{request.get_absolute_url()}"
-        label_fields = [
-            ("📝 Request Template", request.request_template),
-            ("📦 Product Context", request.product_context),
-            ("📌 Applies To", request.content_object),
-            ("🙋 Submitted By", request.requester),
-            ("👤 Assigned To", request.assignee),
-            ("🚨 Priority", request.priority),
-            ("🗒️ Notes", request.notes),
-            ("🔗️ DejaCode URL", request_url),
-        ]
-
-        lines = []
-        for label, value in label_fields:
-            if value:
-                lines.append(f"### {label}\n{value}")
-
-        lines.append("----")
-
-        for question in request.get_serialized_data_as_list():
-            label = question.get("label")
-            value = question.get("value")
-            input_type = question.get("input_type")
-
-            if input_type == "BooleanField":
-                value = "Yes" if str(value).lower() in ("1", "true", "yes") else "No"
-
-            lines.append(f"### {label}\n{value}")
-
-        return "\n\n".join(lines)
+def get_class_for_platform(platform):
+    return {
+        "github": GitHubIntegration,
+        "gitlab": GitLabIntegration,
+        "jira": JiraIntegration,
+    }.get(platform)
